refactor: extract direct boto3 ops into s3op_boto.py
Move all new direct-boto3 helper methods into a separate
S3DirectClient class in s3op_boto.py. This keeps the s3.py
diff small and focused on the feature flag dispatch.
